Windows compatibility for setup.py
For windows scripts without an extension to be executable from the
command line, they need to be installed with setuptools (or its
successor, distribute) rather than with distutils.  The former will
create stub .EXE's that launch python modules, when they are listed as
console_scripts.

In order to get console_scripts to work, all the current executable
scripts needed to be moved into a submodule of zeroinstall and given
an entry point method (called "main()" in this case).

This commit was tested with distribute.